KIS-52 corrijo codigo de op para buscar marco

+++ Shared/utils/coordenada.c
@@ -23,7 +23,7 @@ void *coordenada_armar_stream(Coordenada_tabla* coordenada) {
 }
 ;
 
-void* coordenada_serializar(Coordenada_tabla* coordenada, uint32_t* tamano_mensaje, uint32_t codigo_operacion) {
+void* coordenada_serializar(Coordenada_tabla* coordenada, uint32_t* tamano_mensaje, uint8_t codigo_operacion) {
 	void* stream_coordenada = coordenada_armar_stream(coordenada);
 	uint32_t tamano = sizeof(uint32_t) * 2;
